+++ packages/ohm-js/test/test-ohm-syntax.js
@@ -131,12 +131,12 @@ describe('unicode', test => {
 
   test('categories / properties', t => {
     assertSucceeds(t, m.match('\u01C0', 'unicodeLtmo'), 'dental click is in Ltmo');
-    assertSucceeds(t, m.match('\u01C0', 'unicodeIdStart'), 'dental click in ID_Start');
-    assertSucceeds(t, m.match('\u01C0', 'unicodeIdContinue'), 'dental click in ID_Continue');
-    assertFails(t, m.match('!', 'unicodeIdStart'), '`!` NOT in ID_Start');
-    assertFails(t, m.match('!', 'unicodeIdContinue'), '`!` NOT in ID_Continue');
-    assertFails(t, m.match(' ', 'unicodeIdStart'), 'space NOT in ID_Start');
-    assertFails(t, m.match(' ', 'unicodeIdContinue'), 'space NOT in ID_Continue');
+    assertSucceeds(t, m.match('\u01C0', 'unicodeIDStart'), 'dental click in ID_Start');
+    assertSucceeds(t, m.match('\u01C0', 'unicodeIDContinue'), 'dental click in ID_Continue');
+    assertFails(t, m.match('!', 'unicodeIDStart'), '`!` NOT in ID_Start');
+    assertFails(t, m.match('!', 'unicodeIDContinue'), '`!` NOT in ID_Continue');
+    assertFails(t, m.match(' ', 'unicodeIDStart'), 'space NOT in ID_Start');
+    assertFails(t, m.match(' ', 'unicodeIDContinue'), 'space NOT in ID_Continue');
   });
 
   test('semantic actions', t => {